Added tests to capture deserializer calls from serde types
In order to test that serde_dwarf properly deserializes types, we need
to benchmark against the types that serde supports.

As a first step, this commit implements a Mock Deserializer that
records the deserializer trait functions that are called by a type's
Deserialize impl.

This is similar to the behavior of the `serde_test` crate, which
provides assertions that also track functions calls in serde
impls. However, that crate is focused on testing Deserialize/Serialize
impls by constructing types and comparing them to example structs. We
are interested in the opposite test: Taking a specific type and
determining the deserialize function calls it takes to construct it.