Added tests to compare function calls of deserialize impls
`tests/de.rs` has tests for comparing the function calls that the
"real" serde deserialize impls call compared to what we fake with
serde dwarf.

This builds on the previous work in test/tokens that created  a
MockDeserializer that records which Deserializer methods are called.

`tokens.rs` now servers to test that the function calls that serde uses
have not changed, while `de.rs` checks that the actual serde_dwarf
impl matches.