Added support to deserialize structs and enums
The current deserialization solution only works on non-self-describing
data formats such as bincode, because we expect the info about
selected enum variants to be provided as integers, not as strings.

In addition, there are some weird hacks because serde requires that
struct and enum deserialization provide 'static string names and field
descriptions, which is very hard to generate dynamically. It seems
that a lot of data formats could get away with 'de lifetime string
names, but serde is setup with static lifetimes in order to support
some strange edgecases.

This means for now we provide empty static strings (""), which
technically satisfies the requirements, but will cause problems when
trying to re-serialize into text based formats. I expect JSON will be
a popular format, which of course will need to know the names of
fields in order to complete the serialization.

This leaves a couple more options for future implementation:

1. `unsafe` implementation that uses mem::transmute to give 'de
lifetime variables to the serde Serializer/Deserializer. This could
work for most practical data formats, but is kind of icky since we are
relying on implementation details of those formats.

2. String interning of all type description strings in a global
pool. This approach doesn't violate the requirements of the serde API,
and may in fact have a lot of benefits regarding heap allocation. It
does have the downside, however, that if you are loading in a bunch of
.debuginfo files in the same process, this will cause uncontrolled
growth of memory usage.

It seems like #2 has the most benefits, and shouldnt be too hard to
do. It does rule out applications with long running processes, but
only those that load in a bunch of symbol files over time. A logger /
monitor process still wouldn't have issues since we only grow the
string interning pool with the symbols in the file, and this can't be
much bigger than the mmap'ed source file anyways.