Added Error types to entry_parser and type_builder
This change doesnt completely add error types, but it starts to lay
the groundwork. There is a simple Error impl called InternalError that
can be chained with each error, as they bubble up. However, this error
is not exposed in the API, it is just dumped to stderr at the top part.

This at least provides some helpful info on why the parser is
breaking.