Remove incomplete types from builder
We now properly remove incomplete types from the builder cache. This
means that if we are trying to add a complex type, and one of its
dependent types fails to be added, we remove the parent type from the
builder cache, so that no one can try to deserialize with that type.