Added support for Vec
Added support for Vec. This is still done as a specialization in the
main DIE parser, I will need to fix this soon to reduce the boiler
plate when writing specializations for data types. The Seq Visitor
also has some specializations to make sure that the length of the seq
matches what the type describes.

The last main deserialization type to add is Map, and then after that
I can focus on adding more specialization for the serde data types,
for example, slices.